add controller to auth routes
